--- app/utils/get_ui_components.py
@@ -1,3 +1,0 @@
-def get_ui_components():
-    from ui import transcriber
-    return transcriber
