+++ ui.py
@@ -1,7 +1,7 @@
 import warnings
+import torch
 from tkinter import simpledialog
-from app.core.recorder import AudioRecorder
-from app.gui.components.waveform import WaveformVisualizer
+from recorder import AudioRecorder
 from tkinter import colorchooser
 recorder = AudioRecorder()
 
@@ -14,18 +14,21 @@ warnings.filterwarnings(
 )
 
 import tkinter as tk
-from tkinter import filedialog, messagebox
-from app.core.recorder import AudioRecorder
-from app.core.transcriber import AudioTranscriber
-from app.core.emotion_analyzer import EmotionAnalyzer
-from app.core.text_processor import TextProcessor
-from app.core.text_analyzer import TextAnalyzer
-from app.gui.handlers.export import export_transcription
+from tkinter import filedialog, TclError, messagebox
+from recorder import AudioRecorder
+from transcriber import AudioTranscriber
+from emotion_analyzer import EmotionAnalyzer
+from text_processor import TextProcessor
+from text_analyzer import TextAnalyzer
 import logging
 import warnings
 import os
-from tkinterdnd2 import TkinterDnD
+from tkinterdnd2 import TkinterDnD, DND_FILES
+import threading
 from matplotlib.backends.backend_tkagg import FigureCanvasTkAgg
+from matplotlib.figure import Figure
+import numpy as np
+import pyaudio
 import hashlib
 import collections
 import datetime
@@ -33,6 +36,8 @@ import matplotlib
 matplotlib.use("TkAgg")
 import matplotlib.pyplot as plt
 import seaborn as sns
+import json
+import csv
 # Suppress FP16 warning
 warnings.filterwarnings("ignore", message="FP16 is not supported on CPU; using FP32 instead")
 
@@ -67,34 +72,12 @@ light_theme = {
     'grid_color': '#dddddd'
 }
 
-dark_theme = {
-    'bg': '#2b2b2b',
-    'fg': 'white',
-    'button_bg': '#4caf50',
-    'button_fg': 'white',
-    'text_bg': '#333333',
-    'text_fg': 'white',
-    'plot_bg': '#2b2b2b',
-    'plot_fg': 'white',
-    'waveform_color': '#4caf50',
-    'axis_color': 'white',
-    'grid_color': '#444444'
-}
-
-light_theme = {
-    'bg': '#f0f0f0',
-    'fg': 'black',
-    'button_bg': '#4caf50',
-    'button_fg': 'white',
-    'text_bg': 'white',
-    'text_fg': 'black',
-    'plot_bg': 'white',
-    'plot_fg': 'black',
-    'waveform_color': '#4caf50',
-    'axis_color': 'black',
-    'grid_color': '#dddddd'
-}
 current_theme = dark_theme
+
+def toggle_theme():
+    global current_theme
+    current_theme = light_theme if current_theme == dark_theme else dark_theme
+    apply_theme(current_theme)
 
 def apply_theme(theme):
     # Update root and main frames
@@ -129,12 +112,6 @@ def apply_theme(theme):
     
     # Update seaborn style
     sns.set_style("darkgrid" if theme == dark_theme else "whitegrid")
-
-def toggle_theme():
-    global current_theme
-    current_theme = light_theme if current_theme == dark_theme else dark_theme
-    apply_theme(current_theme)
-
 class TextBoxLogHandler(logging.Handler):
     def __init__(self, text_widget):
         super().__init__()
@@ -150,7 +127,21 @@ class TextBoxLogHandler(logging.Handler):
 # Set default save directory to the current working directory
 save_directory = os.getcwd()
 logging.info(f"Default save directory set to: {save_directory}")
-
+class DropZone(tk.Label):
+    def __init__(self, master, **kwargs):
+        super().__init__(master, **kwargs)
+        self.configure(relief="groove", borderwidth=2)
+        
+        # Enable drag and drop for files
+        self.drop_target_register(DND_FILES)
+        self.dnd_bind('<<Drop>>', self.handle_drop)
+        
+    def handle_drop(self, event):
+        # Get the dropped file path and handle it
+        file_path = event.data
+        if file_path.startswith('{') and file_path.endswith('}'):
+            file_path = file_path[1:-1]
+        handle_dropped_file(file_path)
 def update_font_size(new_size):
     """Update font size for the transcription box"""
     current_font = transcription_box.cget("font")
@@ -529,6 +520,87 @@ def query_text(event=None):
         
     except Exception as e:
         logging.error(f"Error processing query: {e}")
+
+
+class WaveformVisualizer:
+    def __init__(self, frame):
+        self.frame = frame
+        self.is_recording = False
+        
+        # Create matplotlib figure
+        self.fig = Figure(figsize=(4, 4), dpi=100, facecolor='#2b2b2b')
+        self.ax = self.fig.add_subplot(111)
+        self.ax.set_facecolor('#2b2b2b')
+        self.ax.tick_params(axis='x', colors='white')
+        self.ax.tick_params(axis='y', colors='white')
+        
+        # Set up the line plot with matching dimensions
+        self.chunk_size = 1024
+        self.x = np.arange(0, self.chunk_size)
+        self.line, = self.ax.plot(self.x, np.zeros(self.chunk_size), color='#4caf50')
+        
+        # Configure plot appearance
+        self.ax.set_ylim(-32768, 32767)
+        self.ax.set_xlim(0, self.chunk_size)
+        self.ax.grid(True, color='#444444')
+        
+        # Create canvas
+        self.canvas = FigureCanvasTkAgg(self.fig, master=self.frame)
+        self.canvas_widget = self.canvas.get_tk_widget()
+        self.canvas_widget.configure(bg='#2b2b2b', highlightthickness=0)
+        self.canvas_widget.pack(fill='both', expand=True, padx=10, pady=5)
+        
+        # Audio stream configuration
+        self.format = pyaudio.paInt16
+        self.channels = 1
+        self.rate = 44100
+        self.p = None
+        self.stream = None
+
+    def start_recording(self):
+        self.is_recording = True
+        threading.Thread(target=self._record_stream, daemon=True).start()
+
+    def stop_recording(self):
+        self.is_recording = False
+        if self.stream:
+            self.stream.stop_stream()
+            self.stream.close()
+        if self.p:
+            self.p.terminate()
+
+    def _record_stream(self):
+        self.p = pyaudio.PyAudio()
+        self.stream = self.p.open(
+            format=self.format,
+            channels=self.channels,
+            rate=self.rate,
+            input=True,
+            frames_per_buffer=self.chunk_size
+        )
+
+        while self.is_recording:
+            try:
+                data = self.stream.read(self.chunk_size)
+                audio_data = np.frombuffer(data, dtype=np.int16)
+                self.line.set_ydata(audio_data)
+                self.canvas.draw_idle()
+            except Exception as e:
+                print(f"Error reading audio stream: {e}")
+                break
+
+        self.stop_recording()
+    def update_theme(self, theme):
+        self.fig.set_facecolor(theme['plot_bg'])
+        self.ax.set_facecolor(theme['plot_bg'])
+        self.line.set_color(theme['waveform_color'])
+        self.ax.tick_params(axis='x', colors=theme['axis_color'])
+        self.ax.tick_params(axis='y', colors=theme['axis_color'])
+        self.ax.grid(True, color=theme['grid_color'])
+        self.ax.xaxis.label.set_color(theme['axis_color'])
+        self.ax.yaxis.label.set_color(theme['axis_color'])
+        self.ax.title.set_color(theme['axis_color'])
+        self.canvas.draw()
 
 recorder = AudioRecorder()
 transcriber = AudioTranscriber()
